Move ITestHostRunner to appropriate folder
